Match against stacktrace in text filter
Since a stacktrace is an array of frame, we add a function
to test if a string is included in a frame, and we use it
when testing the message's location and its stacktrace - for
each frame of the stack.

Adds some test to ensure we match the function name and the
location in the stacktrace, as well as test to make sur we
don't get unwanted messages.